Support a systemd wpa_supplicant service instance
The `wpa_supplicant.service` unit on Debian supports being called as an
instance for a specific interface, for example:
`wpa_supplicant@wlan0.service`

The service can initialize the specified interface accordingly. It will
look for a configuation file at
`/etc/wpa_supplicant/wpa_supplicant-<iface>.conf` and use that as the
config file for `wpa_supplicant`. At the moment this creates that file
as a symlink to the general `wpa_supplicant.conf` config file, and
enables & starts the service instance.

Useful in combination with networking tools that do not perform AP
association by themselves, such as systemd-networkd.